requirements.txt
flask
kafka-python
redis
pyflink
requests
flask-cors
numpy
scikit-learn
grpcio
grpcio-tools
protobuf